rearanging tests
