+++ services/web/server/tests/unit/with_dbs/01/wallets/test_wallets_groups.py
